+++ spec/pio/open_flow13/match_spec.rb
@@ -645,6 +645,54 @@ describe Pio::OpenFlow13::Match do
       And { match.match_fields[1].masked? == true }
       And { match.match_fields[1].oxm_length == 12 }
     end
+
+    context 'with tunnel_id: 1' do
+      When(:match) do
+        Pio::OpenFlow13::Match.new(tunnel_id: 1)
+      end
+      Then { match.tunnel_id == 1 }
+      And { match.class == Pio::OpenFlow13::Match }
+      And { match.length == 16 }
+      And { match.match_type == Pio::OpenFlow13::MATCH_TYPE_OXM }
+      And { match.match_length == 16 }
+      And { match.match_fields.size == 1 }
+      And do
+        match.match_fields[0].oxm_class ==
+          Pio::OpenFlow13::Match::OXM_CLASS_OPENFLOW_BASIC
+      end
+      And do
+        match.match_fields[0].oxm_field ==
+          Pio::OpenFlow13::Match::TunnelId::OXM_FIELD
+      end
+      And { match.match_fields[0].masked? == false }
+      And { match.match_fields[0].oxm_length == 8 }
+    end
+
+    context 'with tunnel_id: 1, tunnel_id_mask: 0x8000000000000000' do
+      When(:match) do
+        Pio::OpenFlow13::Match.new(
+          tunnel_id: 1,
+          tunnel_id_mask: 0x8000000000000000
+        )
+      end
+      Then { match.tunnel_id == 1 }
+      Then { match.tunnel_id_mask == 0x8000000000000000 }
+      And { match.class == Pio::OpenFlow13::Match }
+      And { match.length == 24 }
+      And { match.match_type == Pio::OpenFlow13::MATCH_TYPE_OXM }
+      And { match.match_length == 24 }
+      And { match.match_fields.size == 1 }
+      And do
+        match.match_fields[0].oxm_class ==
+          Pio::OpenFlow13::Match::OXM_CLASS_OPENFLOW_BASIC
+      end
+      And do
+        match.match_fields[0].oxm_field ==
+          Pio::OpenFlow13::Match::TunnelId::OXM_FIELD
+      end
+      And { match.match_fields[0].masked? == true }
+      And { match.match_fields[0].oxm_length == 16 }
+    end
   end
 
   def read_raw_data_file(name)